rpm-requirements.txt
python-kombu
python-oslo-config
PyYAML
python-taskflow
